Parse each ward, not all wards
This was a subtle bug that led to every councillor being assigned the
same ward. We want to find all `ward` key, not all `wards`. The latter
means we return the object containing all wards and councillors, meaning
when we later call `ward.find("wardtitle").text` it will return the
first match, the one at the start of the `wards` object.