Move councillors scrapers to councillors app
Scrapers for each type of thing that we want to scrape should live in
that thing's directory. This means LGSF/lgsf/scrapers should just be the
resources common to all scrapers. This is being changed ahead of adding
in polling stations as a new 'type of thing' to be scraped.