Configure CircleCI and use python 3.8 everywhere.
SAM config files are now kept in a directory so that there can be one
per environment. Local (as opposed to ci) deploys should use
samconfig.toml.d/development.toml.

At the moment only development deploys will work staging and production
need to be implemented, however this commit lays groundwork for that,
following the pattern in the aggregator-api repository.